# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {
            "name": "Brick Support",
            "url": "https://www.brick.com/support",
            "email": "support@brick.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "paths": {
...
        "/transfer": {
            "get": {
                "security": [
                    {
                        "OAuth2Password": []
                    }
                ],
                "description": "get transfer bank",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transfer"
                ],
                "summary": "get Transfer data",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "search by id",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "search by job id",
                        "name": "job_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "search by api key",
                        "name": "api_key",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "search by status",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "sort result by attributes",
                        "name": "sort_by",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": " ",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": " ",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "must-revalidate",
                            "none"
                        ],
                        "type": "string",
                        "description": "Request Cache Control",
                        "name": "Cache-Control",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.TransferJobsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.TransferJobsResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.TransferJobsResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "APIKey": []
                    }
                ],
                "description": "create transfer bank",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transfer"
                ],
                "summary": "Create Transfer data",
                "parameters": [
                    {
                        "description": "Create Transfer Data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.CreateTransfer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.SingleTransferJobResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.SingleTransferJobResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.SingleTransferJobResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.SingleTransferJobResponse"
                        }
                    }
                }
            }
        },
        "/transfer/{job_id}": {
            "get": {
                "security": [
                    {
                        "OAuth2Password": []
                    }
                ],
                "description": "get transfer data by job id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transfer"
                ],
                "summary": "Get Transfer Data By Job ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "get by job id",
                        "name": "job_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "enum": [
                            "must-revalidate",
                            "none"
                        ],
                        "type": "string",
                        "description": "Request Cache Control",
                        "name": "Cache-Control",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.SingleTransferJobResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.SingleTransferJobResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.SingleTransferJobResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "model.CreateTransfer": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "destination_bank_account": {
                    "type": "string"
                },
                "destination_bank_id": {
                    "type": "integer"
                },
                "source_bank_account": {
                    "type": "string"
                },
                "source_bank_id": {
                    "type": "integer"
                },
                "transaction_time": {
                    "type": "string"
                }
            }
        },
        "model.Pagination": {
            "type": "object",
            "properties": {
                "current_elements": {
                    "type": "integer"
                },
                "current_page": {
                    "type": "integer"
                },
                "sort_by": {
                    "type": "string"
                },
                "total_elements": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
...
        "model.TransferJob": {
            "type": "object",
            "properties": {
                "api_key": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "type": "integer"
                },
                "deleted_at": {
                    "type": "string"
                },
                "deleted_by": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "job_id": {
                    "type": "string"
                },
                "payload": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "updated_by": {
                    "type": "integer"
                }
            }
        },
...
        "response.SingleTransferJobResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/model.TransferJob"
                },
                "message": {
                    "type": "string"
                },
                "status_code": {
                    "type": "integer"
                },
                "transaction_info": {
                    "$ref": "#/definitions/response.TransactionInfo"
                },
                "translation": {
                    "$ref": "#/definitions/response.Translation"
                }
            }
        },
        "response.TransactionInfo": {
            "type": "object",
            "properties": {
                "cause": {
                    "type": "string"
                },
                "error_code": {
                    "type": "integer"
                },
                "request_id": {
                    "type": "string"
                },
                "request_method": {
                    "type": "string"
                },
                "request_uri": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
        "response.TransferJobsResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.TransferJob"
                    }
                },
                "message": {
                    "type": "string"
                },
                "pagination": {
                    "$ref": "#/definitions/model.Pagination"
                },
                "status_code": {
                    "type": "integer"
                },
                "transaction_info": {
                    "$ref": "#/definitions/response.TransactionInfo"
                },
                "translation": {
                    "$ref": "#/definitions/response.Translation"
                }
            }
        },
        "response.Translation": {
            "type": "object",
            "properties": {
                "en": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "APIKey": {
            "description": "Type \"APIKey\" followed by api key",
            "type": "apiKey",
            "name": "API-Key",
            "in": "header"
        },
        "OAuth2Password": {
            "type": "oauth2",
            "flow": "password",
            "tokenUrl": "%s"
        }
    }
}